Modify script to join all cover data
Script joins the two rounds of cover sampling. It outputs
three dataframes into the data folder: plot_cover is the average
cover at each site. This is the average cover of the five subplots
per site. Subplot cover is the cover per species per subplot.
Microplot cover is the cover in the microplots at the lower left
and upper right of each of the subplots.

Of the two rounds of sampling (an early and a late), I only use the
maximum cover estimate for each species.